// src/themes/neutral/accentPalettes.generated.ts
export const palette = {
  "amber": {
    "name": "Amber",
    "light": {
      "0": "#000000",
      "5": "#1c0d00",
      "10": "#291800",
      "15": "#362100",
      "20": "#432b00",
      "25": "#513500",
      "30": "#5f4000",
      "35": "#6e4b00",
      "40": "#7e5600",
      "45": "#8d6100",
      "50": "#9d6c00",
      "55": "#ae7800",
      "60": "#be8400",
      "65": "#cf9000",
      "70": "#dd9e25",
      "75": "#e9ad44",
      "80": "#f4bc5f",
      "85": "#ffcb79",
      "90": "#ffddaa",
      "95": "#ffeed5",
      "100": "#ffffff"
    },
// ...
  },
// ...
} as const;